fix(server): включены глобальные типы Express в tsconfig
