+++ packages/server/package.json
@@ -25,7 +25,7 @@
   "devDependencies": {
     "@types/cookie-parser": "^1.4.9",
     "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
+    "@types/express": "^5.0.3",
     "@types/jest": "^28.1.8",
     "@types/pg": "^8.6.5",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
