+++ packages/server/tsconfig.json
@@ -21,9 +21,8 @@
     "lib": ["es2022", "esnext.asynciterable"],
     "types": ["node", "jest"],
     "baseUrl": ".",
-    "paths": {
-      "*": ["types/*"]
-    },
     "outDir": "dist"
-  }
+  },
+  "include": ["./**/*.ts"],
+  "exclude": ["dist"]
 }
