+++ packages/server/types/User.ts
@@ -1,0 +1,10 @@
+export interface User {
+  id: number;
+  login: string;
+  first_name: string;
+  second_name: string;
+  display_name: string | null;
+  email: string;
+  phone: string;
+  avatar: string | null;
+}
