+++ packages/server/types/express.d.ts
@@ -1,0 +1,12 @@
+import { User } from './User'; // Убедись, что путь правильный
+
+declare global {
+  namespace Express {
+    interface Request {
+      user?: User;
+      cookies: Record<string, string>;
+    }
+  }
+}
+
+export {};
